+++ src/cjs/BezierPath.d.ts
@@ -27,7 +27,7 @@
  * @modified 2023-10-07 Adding the `BezierPath.fromCurve(CubicBezierCurve)` static function.
  * @modified 2025-04-09 Added the `BezierPath.move` method to match the convention – which just calls `translate`.
  * @modified 2025-04-09 Modified the `BezierPath.translate` method: chaning parameter `Vertex` to more generalized `XYCoords`.
- 
+ * @modified 2025-04-14 Class `BezierPath` is now implementing interface `Intersectable`.
  * @version 2.7.0
  *
  * @file BezierPath
@@ -35,6 +35,8 @@
  **/
 import { Bounds } from "./Bounds";
 import { CubicBezierCurve } from "./CubicBezierCurve";
+import { Vector } from "./Vector";
+import { VertTuple } from "./VertTuple";
 import { Vertex } from "./Vertex";
 import { XYCoords, SVGSerializable, UID } from "./interfaces";
 /**
@@ -347,6 +349,26 @@ export declare class BezierPath implements SVGSerializable {
      * @return {Vertex} The perpendicluar vector at the relative path position.
      **/
     getPerpendicular(u: number): Vertex;
+    /**
+     * Get all line intersections with this shape.
+     *
+     * This method returns all intersections (as vertices) with this shape. The returned array of vertices is in no specific order.
+     *
+     * @param {VertTuple} line - The line to find intersections with.
+     * @param {boolean} inVectorBoundsOnly - If set to true only intersecion points on the passed vector are returned (located strictly between start and end vertex).
+     * @returns {Array<Vertex>} - An array of all intersections with the shape's outline.
+     */
+    lineIntersections(line: VertTuple<any>, inVectorBoundsOnly?: boolean): Array<Vertex>;
+    /**
+     * Get all line intersections of this polygon and their tangents along the shape.
+     *
+     * This method returns all intersection tangents (as vectors) with this shape. The returned array of vectors is in no specific order.
+     *
+     * @param line
+     * @param lineIntersectionTangents
+     * @returns
+     */
+    lineIntersectionTangents(line: VertTuple<any>, inVectorBoundsOnly?: boolean): Array<Vector>;
     /**
      * This is a helper function to locate the curve index for a given
      * absolute path position u.